pyproject.toml
[tool.poetry]
name = "nidus"
version = "0.1.0"
description = "Nidus: A configurable multi-agent framework on top of LangGraph"
authors = ["Jeff <jeff@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langgraph = ">=0.2.0"
langchain = ">=0.3.0"
langchain-openai = ">=0.2.0"
langchain-anthropic = ">=0.2.0"
langchain-google-genai = ">=2.0.0"
pydantic = ">=2.0.0"
pyyaml = ">=6.0.0"
psycopg2-binary = ">=2.9.0"
chromadb = ">=0.5.0"
duckduckgo-search = ">=6.0.0"
requests = ">=2.31.0"
python-dotenv = ">=1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
